requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiohttp-retry
aiolimiter
aiosignal
alembic
annotated-types
anyio
APScheduler==3.10.4
attrs
bce-python-sdk
beautifulsoup4
blinker
Brotli
cachetools
cairocffi
CairoSVG
certifi
cffi
charset-normalizer
click
cssselect2
defusedxml
Deprecated
dill
diskcache
distro
duckduckgo-search
Flask==2.3.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist
future
gunicorn==21.2.0
h11
h2
hpack
httpcore
httpx
hyperframe
idna
itsdangerous==2.1.2
jieba
Jinja2==3.1.2
jiter
limits
lxml
Mako
Markdown
markdown-it-py
MarkupSafe==2.1.3
mdurl
multidict
multiprocess
openai==1.60.1
volcengine-python-sdk[ark]
ordered-set
packaging
pillow==10.0.0
prompt_toolkit
propcache
psycopg2-binary==2.9.9
pycparser
pycryptodome
pydantic
pydantic_core
Pygments
PyJWT
python-dotenv==1.0.0
PyYAML
qianfan==0.0.4
requests==2.31.0
rich
shellingham
six
sniffio
socksio
soupsieve
SQLAlchemy==2.0.23
tenacity
tinycss2
tqdm
twilio
typer
typing_extensions
tzlocal
urllib3==2.0.4
wcwidth
webencodings
Werkzeug==2.3.7
wrapt
yarl
boto3==1.36.6
#gradio_client==1.7.2
torch==2.6.0
autogen==0.7.5
torchaudio==2.6.0
numpy==2.2.3
azure-cognitiveservices-speech==1.43.0
pydub
redis
